chore: remove meteor.startup from unpin-message
