+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -6,5 +6,4 @@ import './permalinkPinned';
 import './permalinkStar';
 import './pinMessage';
 import './starMessage';
-import './unpinMessage';
 import './unstarMessage';
